+++ src/server/node/server.mjs
@@ -163,7 +163,7 @@ class WebServer extends EventEmitter {
             if (
                 url.indexOf('__OSC_ASSET__=1') != -1 ||
                 url.indexOf('open-stage-control-client.js.map') != -1 ||
-                url.indexOf('/client/ace/') != -1
+                url.indexOf('/client/') == 0
             ) {
 
                 // osc asset files